+++ packages/core/src/transports/base.ts
@@ -3,7 +3,7 @@ import { BaseExtension } from '../extensions';
 import type { Transport, TransportItem } from './types';
 
 export abstract class BaseTransport extends BaseExtension implements Transport {
-  abstract send(item: TransportItem | TransportItem[]): void | Promise<void>;
+  abstract send(items: TransportItem[]): void | Promise<void>;
 
   getIgnoreUrls(): Array<string | RegExp> {
     return [];