+++ packages/core/src/transports/index.ts
@@ -7,7 +7,9 @@ export { TransportItemType, transportItemTypeToBodyKey } from './const';
 export { registerInitialTransports } from './registerInitial';
 
 export type {
+  BatchExecutorOptions,
   BeforeSendHook,
+  SendFn,
   Transport,
   TransportBody,
   TransportItem,